Resolved  warnings while deploying
