+++ server.js
@@ -13,5 +13,3 @@ connectDB()
     .catch(error => {
         console.error('Database connection failed: ', error);
     });
-
-module.exports = server;
